+++ CHANGELOG.md
@@ -2,6 +2,47 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.7](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.6...v0.1.7) (2022-02-05)
+
+
+### Bug Fixes
+
+* :lipstick: bottom space ([7b4a192](https://github.com/yeeezsh/credit-card-reminder/commit/7b4a192f0863174e553cd5721f3720b026fb10cb))
+
+### [0.1.6](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.5...v0.1.6) (2022-02-05)
+
+
+### Features
+
+* :sparkles: add button ([ba7c41c](https://github.com/yeeezsh/credit-card-reminder/commit/ba7c41c37c06d0e35f6adf07e7df51e27b8c4ffb))
+
+### [0.1.5](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.4...v0.1.5) (2022-02-05)
+
+
+### Features
+
+* :sparkles: random color ([52da92a](https://github.com/yeeezsh/credit-card-reminder/commit/52da92ab779c11ad1cfbedffe397ff7f136af40f))
+* :sparkles: render card ([2bbd6f4](https://github.com/yeeezsh/credit-card-reminder/commit/2bbd6f46c220a491d38935ef049cc81ee22de4d0))
+
+
+### Bug Fixes
+
+* :see_no_evil: wrong title name ([d2a6b32](https://github.com/yeeezsh/credit-card-reminder/commit/d2a6b32015b282732b2fe78515a3ad13fe0b4308))
+
+### [0.1.4](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.1...v0.1.4) (2022-02-05)
+
+
+### Features
+
+* :sparkles: add main layout ([652abe1](https://github.com/yeeezsh/credit-card-reminder/commit/652abe1e9f1eb802565b4e684c8d942a7d92d416))
+* :sparkles: add mock card ([a5c10df](https://github.com/yeeezsh/credit-card-reminder/commit/a5c10df392c7bd2c8d5bda414023c26056c85902))
+
+
+### Bug Fixes
+
+* :sparkles: add responsive layout for card ([135e514](https://github.com/yeeezsh/credit-card-reminder/commit/135e51452f5853107619f36685ba2265acf33f4a))
+* :wrench: turn private to false ([d59882c](https://github.com/yeeezsh/credit-card-reminder/commit/d59882cbdd5a6697168c822cf063e6a8b210390d))
+
 ### [0.1.3](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.2...v0.1.3) (2022-02-05)
 
 ### [0.1.2](https://github.com/yeeezsh/credit-card-reminder/compare/v0.1.1...v0.1.2) (2022-02-05)